Refactor: Rename CustomerRepository to AuthRepository and update usages
This commit renames `CustomerRepository` to `AuthRepository` to better reflect its purpose.

It also updates the `NetworkModule` to provide `UserService` and `TokenService` instances.

The `LoginScreenViewModel` and `RegisterScreenViewModel` are updated to use the renamed `AuthRepository`.

Additionally, the `RegisterScreen` and `LoginScreen` are updated to use `userState.user?.identifier` instead of `userState.user?.accessCode`. The base URL in `NetworkModule` has also been corrected.