Refactor: Update "Coming Soon" button behavior in Hero Carousel
This commit modifies the "Coming Soon" button in the `MovieHeroSectionCarousel`.

Previously, the "Coming Soon" button was non-interactive. Now, clicking it will navigate to the "More Info" screen for the respective movie.

This change is implemented in `CarouselItemForeground.kt` by passing a `goToMoreInfo` lambda to the `CustomFillButton`. A descriptive text "Click for more info" has also been added next to the button.

Additionally, commented-out code for a gradient overlay in `MovieHeroSectionCarousel.kt` has been removed.