Refactor: Prioritize identifierOrEmail in AuthScreen navigation
This commit updates the logic in `AuthScreen.kt` for determining the identifier used when navigating to login.

Previously, `registrationCode` or `loginRequestCode` would be used if `identifierOrEmail` was blank.

Now, `registrationCode` is used only if `identifierOrEmail` and `loginRequestCode` are both blank. Similarly, `loginRequestCode` is used only if `identifierOrEmail` and `registrationCode` are both blank. This gives precedence to `identifierOrEmail` if it's available.